Move `IsolatedProject` to `org.gradle.api.project` package
