+++ subprojects/core/src/main/java/org/gradle/api/internal/project/DefaultIsolatedProject.java
@@ -16,7 +16,7 @@
 
 package org.gradle.api.internal.project;
 
-import org.gradle.api.IsolatedProject;
+import org.gradle.api.project.IsolatedProject;
 import org.gradle.api.file.Directory;
 
 public final class DefaultIsolatedProject implements IsolatedProject {
